:art: moved all the functios to seperate files and added some controllers
